Add support for file names containing spaces in file references.
Fixed the GetText tools to properly support file references when the
file name contains spaces. In addition, modified the GetText tools so
that they will no longer combine multiple file references on a single
line unless the --combine-refs command line argument is specified. This
is to resolve an ambiguity when file names contain spaces.

There is no problem when the file references include line numbers, which
are optional. The function I wrote to find the end of the file name,
po_parse_comment_find_end_of_file_name, is quite capable of dealing with
that case. When the file references contain line numbers, the parsing
code supports multiple references on one line and they can be separated
by either a space or a tab character. However, when line numbers are
left out things become more complicated.

Consider the following example.

#: File Name 1.ext File Name 2.ext

This is only supported if the file references are separated by the tab
character.

To reduce the likelihood of this becoming a problem I chose to have
msgcat and the related tools not combine multiple file references by
default.