Disabled the "warning: internationalized messages should not contain the '{seq}' escape sequence" message.
Disabled the "warning: internationalized messages should not contain the
'{seq}' escape sequence" message for all escape sequences. In JAWS we
make extensive use of these escape sequences and we will continue to do
so. These warnings are causing our localizers to miss errors. There has
already been one incident in which this lead to a malformed PO file
being checked in.

After discussing this with Glen, we decided to simply disable this
warning for all escape sequences.